Remove old text layout/render code
