+++ packages/blitz/src/lib.rs
@@ -8,7 +8,6 @@ mod devtools;
 mod fontcache;
 mod imagecache;
 pub mod render;
-mod text;
 mod util;
 mod viewport;
